+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Version 1.2.6 (Nov 30, 2021 JST)
 
-- Make the `readByt`e and `writeByte` methods public instead of private for allow external programs to read and write memory via the CPU (synchronously). 
+- Make the `readByte` and `writeByte` methods public instead of private for allow external programs to read and write memory via the CPU (synchronously). 
 
 ## Version 1.2.5 (Oct 23, 2021 JST)