+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version 1.2.6 (Nov 30, 2021 JST)
 
 - Make the `readByte` and `writeByte` methods public instead of private for allow external programs to read and write memory via the CPU (synchronously). 
+- Add the ability to handle CALL/RET
 
 ## Version 1.2.5 (Oct 23, 2021 JST)